feat: move user model and user authentication response to shared module
