+++ app/src/main/java/com/singaludra/gofood/feature/login/http/LoginUserHttpClient.kt
@@ -3,7 +3,7 @@ package com.singaludra.gofood.feature.login.http
 import com.singaludra.gofood.feature.login.domain.LoginData
 import com.singaludra.gofood.feature.login.http.model.LoginDataRequest
 import com.singaludra.gofood.feature.register.http.HttpClientRegisterResult
-import com.singaludra.gofood.feature.register.http.response.RemoteRootRegisterUser
+import com.singaludra.gofood.shared.http.response.RemoteRootRegisterUser
 import kotlinx.coroutines.flow.Flow
 
 sealed class HttpClientLoginResult {
