+++ app/src/main/java/com/singaludra/gofood/feature/login/http/LoginUserService.kt
@@ -1,7 +1,7 @@
 package com.singaludra.gofood.feature.login.http
 
 import com.singaludra.gofood.feature.login.http.model.LoginDataRequest
-import com.singaludra.gofood.feature.register.http.response.RemoteRootRegisterUser
+import com.singaludra.gofood.shared.http.response.RemoteRootRegisterUser
 import retrofit2.http.Body
 import retrofit2.http.POST
